pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "research-agent"
version = "0.1.0"
description = "Deep research AI system for generating publication-ready LaTeX papers"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Research Agent Team" }
]
keywords = ["research", "ai", "llm", "latex", "academic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Web framework
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "python-multipart==0.0.20",
    "sse-starlette==2.2.1",

    # Data validation and settings
    "pydantic==2.10.4",
    "pydantic-settings==2.7.1",

    # HTTP client
    "httpx==0.28.1",

    # Agent orchestration
    "langgraph==1.0.6",
    "langgraph-checkpoint-postgres==3.0.3",
    "psycopg[binary,pool]==3.2.4",

    # LLM APIs
    "openai==1.62.0",

    # Search APIs
    "tavily-python==0.5.0",

    # ArXiv integration
    "arxiv==2.1.3",

    # LaTeX and output generation
    "jinja2==3.1.5",
    "markdown==3.7",
    "bibtexparser==2.0.0b7",
    "pylatexenc==2.10",

    # Logging and monitoring
    "structlog==24.4.0",
    "opentelemetry-api==1.29.0",
    "opentelemetry-sdk==1.29.0",
    "opentelemetry-instrumentation-fastapi==0.50b0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.2",
    "pytest-cov==6.0.0",
    "ruff==0.9.2",
    "mypy==1.14.1",
    "pre-commit==4.0.1",
    "pip-audit==2.8.0",
]

brain = [
    # Only needed for running the brain service
    "vllm==0.13.0",
    "transformers==4.48.0",
    "torch==2.5.1",
]

[project.scripts]
research-agent = "src.main:main"

[project.urls]
Homepage = "https://github.com/research-agent/research-agent"
Documentation = "https://github.com/research-agent/research-agent#readme"
Repository = "https://github.com/research-agent/research-agent"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "S",      # flake8-bandit (security)
    "T20",    # flake8-print
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "S101",   # assert used (needed for tests)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101", "ARG", "S104"]  # S104: 0.0.0.0 in test assertions
"src/config.py" = ["S104"]  # 0.0.0.0 binding intentional for Docker

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
exclude = ["tests/"]

[[tool.mypy.overrides]]
module = ["tavily.*", "arxiv.*", "langgraph.*", "sse_starlette.*", "bibtexparser.*", "pylatexenc.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
